+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.7
 
 * Add changelog and features
+* Rename command from `kr8` -> `kr8p`
 
 ## 0.0.6